+++ apicast/src/threescale_utils.lua
@@ -1,9 +1,13 @@
+local sub = string.sub
+local tonumber = tonumber
 
 local redis = require 'resty.redis'
 local env = require 'resty.env'
 
 local resty_resolver = require 'resty.resolver'
 local resty_balancer = require 'resty.balancer'
+
+local resty_url = require 'resty.url'
 
 local _M = {} -- public interface
 
@@ -116,17 +120,57 @@ function _M.resolve(host, port)
   return ip, port
 end
 
-function _M.connect_redis(host, port)
-  local h = host or env.get('REDIS_HOST') or "127.0.0.1"
-  local p = port or env.get('REDIS_PORT') or 6379
+function _M.connect_redis(options)
+  local opts = {}
+
+  local url = options and options.url or env.get('REDIS_URL')
+
+
+  if url then
+    url = resty_url.split(url, 'redis')
+    if url then
+      opts.host = url[4]
+      opts.port = url[5]
+      opts.db = url[6] and tonumber(sub(url[6], 2))
+      opts.password = url[3] or url[2]
+    end
+  elseif options then
+    opts.host = options.host
+    opts.port = options.port
+    opts.db = options.db
+    opts.password = options.password
+  end
+
+  opts.timeout = options and options.timeout or redis_conf.timeout
+
+  local host = opts.host or env.get('REDIS_HOST') or "127.0.0.1"
+  local port = opts.prot or env.get('REDIS_PORT') or 6379
+
   local red = redis:new()
 
-  red:set_timeout(redis_conf.timeout)
+  red:set_timeout(opts.timeout)
 
-  local ok, err = red:connect(_M.resolve(h, p))
+  local ok, err = red:connect(_M.resolve(host, port))
   if not ok then
-    return nil, _M.error("failed to connect to redis on " .. h .. ":" .. p .. ":", err)
+    return nil, _M.error("failed to connect to redis on " .. host .. ":" .. port .. ":", err)
   end
+
+  if opts.password then
+    ok = red:auth(opts.password)
+
+    if not ok then
+      return nil, _M.error("failed to auth on redis " .. host .. ":" .. port)
+    end
+  end
+
+  if opts.db then
+    ok = red:select(opts.db)
+
+    if not ok then
+      return nil, _M.error("failed to select db " .. opts.db .. " on redis " .. host .. ":" .. port)
+    end
+  end
+
   return red
 end
 
@@ -137,9 +181,13 @@ end
 
 -- error and exist
 function _M.error(...)
-  ngx.status = ngx.HTTP_INTERNAL_SERVER_ERROR
-  ngx.say(...)
-  ngx.exit(ngx.status)
+  if ngx.get_phase() == 'timer' then
+    return ...
+  else
+    ngx.status = ngx.HTTP_INTERNAL_SERVER_ERROR
+    ngx.say(...)
+    ngx.exit(ngx.status)
+  end
 end
 
 function _M.missing_args(text)